+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 1.74 (not yet released)
 ------------
 * Add case_id to the GetActivity action.
 * Extended Create/Update Individual with Is Deceased and Deceaded Date.
+* Fixed issue with custom fields and Participant: register contact for event action.
 
 Version 1.73
 ------------